drop TV functionality
