+++ src/parameterized.jl
@@ -132,3 +132,76 @@ asparams(::Type{<:SnedecorF}, ::StaticSymbol{:ν1}) = asℝ₊
 asparams(::Type{<:SnedecorF}, ::StaticSymbol{:ν2}) = asℝ₊
 
 asparams(::Type{<:StudentT}, ::StaticSymbol{:ν}) = asℝ₊
+
+function as(d::PowerMeasure)
+    as(Array, as(d.parent), length.(d.axes)...)
+end
+
+function as(d::ProductMeasure{<:AbstractArray{<:Dirac}})
+    return asConst(testvalue.(marginals(d)))
+end
+
+function as(d::ProductMeasure{A}) where {A<:AbstractArray}
+    mar = marginals(d)
+    ts = map(as, mar)
+    if allequal(ts)
+        return as(Array, first(ts), size(ts))
+    else
+        error("Not yet implemented")
+    end
+end
+
+
+function as(d::ProductMeasure{A}) where {A<:MappedArrays.ReadonlyMappedArray}
+    d1 = marginals(d).f(first(marginals(d).data))
+    as(Array, as(d1), size(marginals(d))...)
+end
+
+function as(d::ProductMeasure{T}) where {T<:Tuple}
+    as(map(as, d.marginals))
+end
+
+function as(d::ProductMeasure{<:Base.Generator})
+    d1 = marginals(d).f(first(marginals(d).iter))
+    as(Array, as(d1), size(marginals(d))...)
+end
+
+as(args...; kwargs...) = TV.as(args...; kwargs...)
+
+as(::Beta) = as𝕀
+as(::Cauchy) = asℝ
+as(d::Dirichlet{(:α,)}) = TV.UnitSimplex(length(d.α))
+as(::Exponential) = asℝ₊
+as(::Gamma) = asℝ₊
+as(::Gumbel) = asℝ
+as(::InverseGaussian) = asℝ₊
+as(::Laplace) = asℝ
+as(d::MvNormal{(:μ,)}) = as(Array, length(d.μ))
+
+as(d::MvNormal{(:Σ,),Tuple{C}}) where {C<:Cholesky} = as(Array, size(d.Σ, 1))
+as(d::MvNormal{(:Λ,),Tuple{C}}) where {C<:Cholesky} = as(Array, size(d.Λ, 1))
+as(d::MvNormal{(:μ, :Σ),<:Tuple{T,C}}) where {T,C<:Cholesky} = as(Array, size(d.Σ, 1))
+as(d::MvNormal{(:μ, :Λ),<:Tuple{T,C}}) where {T,C<:Cholesky} = as(Array, size(d.Λ, 1))
+
+function as(d::MvNormal{(:σ,),Tuple{M}}) where {M<:Triangular}
+    σ = d.σ
+    if @inbounds all(i -> σ[i] ≠ 0, diagind(σ))
+        return as(Array, size(σ, 1))
+    else
+        @error "Not implemented yet"
+    end
+end
+
+function as(d::MvNormal{(:λ,),Tuple{M}}) where {M<:Triangular}
+    λ = d.λ
+    if @inbounds all(i -> λ[i] > 0, diagind(λ))
+        return as(Array, size(λ, 1))
+    else
+        @error "Not implemented yet"
+    end
+end
+
+as(::Normal) = asℝ
+xform(::SnedecorF) = asℝ₊
+xform(::StudentT) = asℝ
+xform(::Uniform{()}) = as𝕀
