+++ src/parameterized/beta.jl
@@ -13,8 +13,6 @@ export Beta
     beta => β
 ]
 
-as(::Beta) = as𝕀
-
 @inline function logdensity_def(d::Beta{(:α, :β),Tuple{A,B}}, x::X) where {A,B,X}
     return xlogy(d.α - 1, x) + xlog1py(d.β - 1, -x)
 end
