+++ src/parameterized/gamma.jl
@@ -42,7 +42,6 @@ Base.rand(rng::AbstractRNG, T::Type, μ::Gamma{()}) = rand(rng, T, Exponential()
 
 Base.rand(rng::AbstractRNG, T::Type, μ::Gamma{(:k,)}) = rand(rng, Dists.Gamma(μ.k))
 
-as(::Gamma) = asℝ₊
 
 insupport(::Gamma, x) = x > 0
