+++ src/parameterized/laplace.jl
@@ -38,4 +38,3 @@ Base.rand(rng::AbstractRNG, ::Type{T}, μ::Laplace) where {T} = Base.rand(rng, T
 
 ≪(::Laplace, ::Lebesgue{X}) where {X<:Real} = true
 
-as(::Laplace) = asℝ
